requirements.txt
autopep8==1.7.0
cachetools==5.2.0
certifi==2022.9.14
chardet==5.0.0
charset-normalizer==2.1.1
colorama==0.4.5
cssselect==1.1.0
cssutils==2.6.0
DateTime==4.7
emails==0.6
et-xmlfile==1.1.0
idna==3.4
lxml==4.9.1
openpyxl==3.0.10
Pillow==9.2.0
premailer==3.10.0
pycodestyle==2.9.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
six==1.16.0
termcolor==2.0.1
toml==0.10.2
urllib3==1.26.12
zope.interface==5.4.0
